Financial Portfolio Tracker
Topic Description:
Our selected topic for this project is, “Design a Stock Portfolio Tracker using
Python”. For this project, we used Tkinter to design and implement our basic
graphical user interfaces. Tkinter helped us to incorporate widgets and event-driven
programming principles to enhance user interaction and experience. We decided on this topic because the features that we wanted to implement in our
program would challenge us to put into practice what we have learned over the course
of the MH6803 module.
The main features of this program included:
1. User Interface & Input Handling
2. MarketStack API Integration:
3. Portfolio Management & Database Integration
4. Visualization & Historical Data Analysis
The program design, programming codes, and final implementation of the user
interface allowed us to comprehensively experiment, explore, and troubleshoot the
various python codes that were used in the delivery of the final product.